+++ tools/maintenance/list_bep_leads.py
@@ -1,8 +1,9 @@
 """List the emails of the leader of each BEP."""
 
 import pandas as pd
-from bids_website.utils import data_dir
+from bids_website.utils import bids_spec_dir, data_dir
 from ruamel.yaml import YAML
+from utils import load_citation, return_lead_from_citation_cff
 
 beps = data_dir() / "beps" / "beps.yml"
 
@@ -10,13 +11,25 @@ with open(beps, "r") as f:
     yaml = YAML(typ="safe", pure=True)
     data = yaml.load(f)
 
+citation = load_citation(bids_spec_dir() / "CITATION.cff")
+
 output = {"bep": [], "lead": [], "email": []}
 
 for bep in data:
     for lead in bep["leads"]:
-        output["bep"].append(f'{bep["number"]} - {bep["title"]}')
-        output["lead"].append(f'{lead["given-names"]} {lead["family-names"]}')
-        output["email"].append(f'{lead["email"]}')
+        if not lead["family-names"].strip():
+            continue
+
+        contributor = return_lead_from_citation_cff(citation, lead)
+        if contributor is None:
+            continue
+
+        if "email" in contributor:
+            output["bep"].append(f'{bep["number"]} - {bep["title"]}')
+            output["lead"].append(
+                f'{lead["given-names"]} {lead["family-names"]}'
+            )
+            output["email"].append(f'{contributor["email"]}')
 
 df = pd.DataFrame(output)
